vision/requirements.txt
Django==4.2.27
Pillow==10.3.0
openai==1.30.1
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==22.0.0
prometheus-client==0.19.0
supabase==2.4.6
requests==2.32.3
ultralytics==8.2.0
